requirements.txt
asgiref==3.6.0
build==0.10.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
docopt==0.6.2
idna==3.4
mysqlclient==2.1.1
packaging==23.0
Pillow==9.4.0
pip-tools==6.12.3
pipreqs==0.4.11
pyproject_hooks==1.0.0
requests==2.28.2
sqlparse==0.4.3
total==0.0.7
tzdata==2022.7
urllib3==1.26.15
yarg==0.1.9
django-debug-toolbar~=3.8.1
django-simple-captcha